--- Client/src/components/Day/Day.jsx
@@ -1,9 +1,0 @@
-const Day = () => {
-  return (
-    <div>
-      Day
-    </div>
-  )
-};
-
-export default Day
