--- Client/src/components/Year/Year.jsx
@@ -1,9 +1,0 @@
-const Year = () => {
-  return (
-    <div>
-      Year
-    </div>
-  )
-};
-
-export default Year
