--- Client/src/views/home/home.jsx
@@ -1,10 +1,0 @@
-import Month from "../../components/Calendar/Calendar";
-
-const HomePage = () => {
-  return (
-    <>
-      <Month />
-    </>
-  );
-};
-export default HomePage;
