+++ Client/src/views/month/month.jsx
@@ -1,0 +1,10 @@
+import MCalendar from "../../components/MCalendar/MCalendar";
+
+const Month = () => {
+  return (
+    <div>
+      <MCalendar />
+    </div>
+  )
+}
+export default Month;
